WIp: refactor
