+++ packages/backend/src/services/UnfurlService/UnfurlService.ts
@@ -1,0 +1,330 @@
+import { assertUnreachable, AuthorizationError } from '@lightdash/common';
+import fetch from 'node-fetch';
+import { LightdashConfig } from '../../config/parseConfig';
+import Logger from '../../logger';
+import { DashboardModel } from '../../models/DashboardModel/DashboardModel';
+import { SavedChartModel } from '../../models/SavedChartModel';
+import { ShareModel } from '../../models/ShareModel';
+import { SpaceModel } from '../../models/SpaceModel';
+import { getAuthenticationToken } from '../../routers/headlessBrowser';
+import { EncryptionService } from '../EncryptionService/EncryptionService';
+
+const puppeteer = require('puppeteer');
+
+const uuid = '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}';
+const uuidRegex = new RegExp(uuid, 'g');
+const nanoid = '[\\w-]{21}';
+const nanoidRegex = new RegExp(nanoid);
+
+export enum LightdashPage {
+    DASHBOARD = 'dashboard',
+    CHART = 'chart',
+    EXPLORE = 'explore',
+}
+
+export type Unfurl = {
+    title: string;
+    description?: string;
+    imageUrl: string;
+    pageType: LightdashPage;
+};
+
+export type ParsedUrl = {
+    isValid: boolean;
+    lightdashPage?: LightdashPage;
+    url: string;
+    dashboardUuid?: string;
+    projectUuid?: string;
+    chartUuid?: string;
+    exploreModel?: string;
+};
+
+type UnfurlServiceDependencies = {
+    lightdashConfig: LightdashConfig;
+    dashboardModel: DashboardModel;
+    savedChartModel: SavedChartModel;
+    spaceModel: SpaceModel;
+    shareModel: ShareModel;
+    encryptionService: EncryptionService;
+};
+
+const saveScreenshot = async (
+    imageId: string,
+    cookie: string,
+    parsedUrl: ParsedUrl,
+): Promise<Buffer> => {
+    let browser;
+
+    try {
+        const browserWSEndpoint = `ws://${process.env.HEADLESS_BROWSER_HOST}:${process.env.HEADLESS_BROWSER_PORT}`;
+        browser = await puppeteer.connect({
+            browserWSEndpoint,
+        });
+
+        const page = await browser.newPage();
+
+        await page.setExtraHTTPHeaders({ cookie });
+
+        await page.setViewport({
+            width: 1400,
+            height: 768, // hardcoded
+        });
+
+        const blockedUrls = [
+            'headwayapp.co',
+            'rudderlabs.com',
+            'analytics.lightdash.com',
+            'cohere.so',
+            'intercom.io',
+        ];
+        await page.setRequestInterception(true);
+        page.on('request', (request: any) => {
+            const requestUrl = request.url();
+            if (blockedUrls.includes(requestUrl)) {
+                request.abort();
+                return;
+            }
+
+            request.continue();
+        });
+        await page.goto(parsedUrl.url, {
+            timeout: 100000,
+            waitUntil: 'networkidle0',
+        });
+        const path = `/tmp/${imageId}.png`;
+
+        const selector =
+            parsedUrl.lightdashPage === LightdashPage.DASHBOARD
+                ? '.react-grid-layout'
+                : `.echarts-for-react, [data-testid="visualization"]`; // Get .echarts-for-react, otherwise fallsback to data-testid (tables and bignumbers)
+        await page.waitForSelector(selector);
+        const element = await page.$(selector);
+
+        if (parsedUrl.lightdashPage === LightdashPage.DASHBOARD) {
+            // Remove navbar from screenshot
+            await page.evaluate((sel: any) => {
+                // @ts-ignore
+                const elements = document.querySelectorAll(sel);
+                elements.forEach((el) => el.parentNode.removeChild(el));
+            }, '.bp4-navbar');
+        }
+
+        const imageBuffer = await element.screenshot({
+            path,
+        });
+
+        return imageBuffer;
+    } catch (e) {
+        Logger.error(`Unable to fetch screenshots from headless chromeo ${e}`);
+        return e;
+    } finally {
+        if (browser) await browser.close();
+    }
+};
+
+const getUserCookie = async (userUuid: string): Promise<string> => {
+    const token = getAuthenticationToken(userUuid);
+
+    const response = await fetch(
+        `${process.env.SITE_URL}/api/v1/headless-browser/login/${userUuid}`,
+        {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({ token }),
+        },
+    );
+    if (response.status !== 200) {
+        throw new Error(
+            `Unable to get cookie for user ${userUuid}: ${await response.text()}`,
+        );
+    }
+    const header = response.headers.get('set-cookie');
+    if (header === null) {
+        const loginBody = await response.json();
+        throw new AuthorizationError(
+            `Cannot sign in:\n${JSON.stringify(loginBody)}`,
+        );
+    }
+    return header;
+};
+
+export class UnfurlService {
+    lightdashConfig: LightdashConfig;
+
+    dashboardModel: DashboardModel;
+
+    savedChartModel: SavedChartModel;
+
+    spaceModel: SpaceModel;
+
+    shareModel: ShareModel;
+
+    encryptionService: EncryptionService;
+
+    constructor({
+        lightdashConfig,
+        dashboardModel,
+        savedChartModel,
+        spaceModel,
+        shareModel,
+        encryptionService,
+    }: UnfurlServiceDependencies) {
+        this.lightdashConfig = lightdashConfig;
+        this.dashboardModel = dashboardModel;
+        this.savedChartModel = savedChartModel;
+        this.spaceModel = spaceModel;
+        this.shareModel = shareModel;
+        this.encryptionService = encryptionService;
+    }
+
+    private async getSharedUrl(linkUrl: string): Promise<string> {
+        const [shareId] = linkUrl.match(nanoidRegex) || [];
+        const shareUrl = await this.shareModel.getSharedUrl(shareId);
+
+        const fullUrl = `${this.lightdashConfig.siteUrl}${shareUrl.path}${shareUrl.params}`;
+        Logger.debug(`Shared url ${shareId}: ${fullUrl}`);
+
+        return fullUrl;
+    }
+
+    private async parseUrl(linkUrl: string): Promise<ParsedUrl> {
+        if (!linkUrl.startsWith(this.lightdashConfig.siteUrl)) {
+            Logger.debug(
+                `URL to unfurl ${linkUrl} does not belong to this siteUrl ${this.lightdashConfig.siteUrl}, ignoring.`,
+            );
+            return {
+                isValid: false,
+                url: linkUrl,
+            };
+        }
+
+        const shareUrl = new RegExp(`/share/${nanoid}`);
+        const url = linkUrl.match(shareUrl)
+            ? await this.getSharedUrl(linkUrl)
+            : linkUrl;
+
+        const dashboardUrl = new RegExp(`/projects/${uuid}/dashboards/${uuid}`);
+        const chartUrl = new RegExp(`/projects/${uuid}/saved/${uuid}`);
+        const exploreUrl = new RegExp(`/projects/${uuid}/tables/`);
+
+        if (url.match(dashboardUrl) !== null) {
+            const [projectUuid, dashboardUuid] =
+                (await url.match(uuidRegex)) || [];
+
+            return {
+                isValid: true,
+                lightdashPage: LightdashPage.DASHBOARD,
+                url,
+                projectUuid,
+                dashboardUuid,
+            };
+        }
+        if (url.match(chartUrl) !== null) {
+            const [projectUuid, chartUuid] = (await url.match(uuidRegex)) || [];
+
+            return {
+                isValid: true,
+                lightdashPage: LightdashPage.CHART,
+                url,
+                projectUuid,
+                chartUuid,
+            };
+        }
+        if (url.match(exploreUrl) !== null) {
+            const [projectUuid] = (await url.match(uuidRegex)) || [];
+
+            const urlWithoutParams = url.split('?')[0];
+            const exploreModel = urlWithoutParams.split('/tables/')[1];
+
+            return {
+                isValid: true,
+                lightdashPage: LightdashPage.EXPLORE,
+                url,
+                projectUuid,
+                exploreModel,
+            };
+        }
+
+        Logger.debug(`URL to unfurl ${url} is not valid`);
+        return {
+            isValid: false,
+            url,
+        };
+    }
+
+    async getTitleAndDescription(
+        parsedUrl: ParsedUrl,
+    ): Promise<{ title: string; description?: string }> {
+        switch (parsedUrl.lightdashPage) {
+            case LightdashPage.DASHBOARD:
+                if (!parsedUrl.dashboardUuid)
+                    throw new Error(
+                        `Missing dashboardUuid when unfurling Dashboard URL ${parsedUrl.url}`,
+                    );
+                const dashboard = await this.dashboardModel.getById(
+                    parsedUrl.dashboardUuid,
+                );
+                return {
+                    title: dashboard.name,
+                    description: dashboard.description,
+                };
+            case LightdashPage.CHART:
+                if (!parsedUrl.chartUuid)
+                    throw new Error(
+                        `Missing chartUuid when unfurling Dashboard URL ${parsedUrl.url}`,
+                    );
+                const chart = await this.savedChartModel.get(
+                    parsedUrl.chartUuid,
+                );
+                return { title: chart.name, description: chart.description };
+            case LightdashPage.EXPLORE:
+                const exploreName = parsedUrl.exploreModel
+                    ? `Exploring ${parsedUrl.exploreModel}`
+                    : 'Explore';
+                return { title: exploreName };
+            case undefined:
+                throw new Error(`Unrecognized page for URL ${parsedUrl.url}`);
+            default:
+                return assertUnreachable(
+                    parsedUrl.lightdashPage,
+                    `No lightdash page Slack unfurl implemented`,
+                );
+        }
+    }
+
+    async unfurl(
+        originUrl: string,
+        imageId: string,
+        authUserUuid: string,
+    ): Promise<Unfurl | undefined> {
+        const parsedUrl = await this.parseUrl(originUrl);
+
+        if (
+            !parsedUrl.isValid ||
+            parsedUrl.lightdashPage === undefined ||
+            parsedUrl.url === undefined
+        ) {
+            return undefined;
+        }
+        Logger.debug(`Unfurling URL ${parsedUrl.url}`);
+
+        const cookie = await getUserCookie(authUserUuid);
+
+        await saveScreenshot(imageId, cookie, parsedUrl);
+
+        const imageUrl = `${this.lightdashConfig.siteUrl}/api/v1/slack/image/${imageId}.png`;
+
+        const { title, description } = await this.getTitleAndDescription(
+            parsedUrl,
+        );
+
+        return {
+            title,
+            description,
+            pageType: parsedUrl.lightdashPage,
+            imageUrl,
+        };
+    }
+}
